fix background Exception on startup
temporary (smile here!) HACK to avoid exception on startup. Android O only allows
Services while showing a notification this shows a notification to satisfy
android. if there are no downloads the service will quite right away and if
there are downloads it will be replaced by a proper notification for the new
episode in queue