bugfix: prevent crash with IndexOutOfBoundException
The app can crash if you drag&drop in the playlist while something
else changes the playlist (e.g. an episode finished).
I don't see an easy fix as the "move" parameters are comming from an
android class. As this is an unlikely case, "handle" it with an ugly
toast-message, because it is still better than a crash.